postprocess.py gets fleshed out, and incrontab tweaks
postprocess.py now provides hardlinked timestamped
directories, and maintains a symlink pointing to
the latest transferred directory.

The timestamped directories are maintained with a resolution
of the nearest minute - we don't expect to have more
than one per hour.

The incrond crontab gets updated to deal well with rsync
clients too. To work well with rsync, the rsync client must
use the -T option to avoid creating tempfiles in the monitored
directory.

Strangely, incrond cannot handle comments on its tab files, so
we workaround that for the moment. This is known as bug 173 in the
incrond bugtracker (bts.aiken.cz).