postprocess.py gets _actually_ fleshed out- and incrontab tweaks
Take two on the patch posted earlier. 2 of the 3 files got dropped.

To recap

postprocess.py now provides hardlined, timestamped
directories, and maintains a symlink pointing to
the latest transferred directory.

the timestamped directories are maintainted with a
resolution of the nearest minute - we don't expect to
have more than one per hour.

The incrond crontab gets updated to deal with rsync
clients too. To work well with rsync, the rsync client
must use the -T option to avoid creating tempfiles in the
monitored directory.

Strangely, incrond cannot handle comments on its tab file
so we put those comments in a separate file (which got committed
on its own, accidentally, earlier).