ds-backup packaging - use git describe for versioning.
Make the rpm versioning dependent on git describe, which does
a very sensible thing - finds the closest tag. If the match
is perfect, it will use the version there stripping a possible
leading 'v' and replacing - with . (so v1.0-33 becomes v1.0.33).

If the match is not perfect, it will count commits of distance,
so 2.0.45.<sha1> is 45 commits after v2.0, with a particular sha1.

It retains the 'snapshot' build number counter, which is a handy
shortcut while testing.