backup-available.py: a more sensible "disk full" refusal
Replace a buggy test (was checking free blocks and file entries
for the _root_ user) with a more sensible test: refuse to backup
if we are over 90% usage on the partition.

ds-cleanup.py should manage the quota - this is a trick to stem
the tide if disk usage is going overboard.