Introducing ds-cleanup - free up space cleaning up old snapshots
This is a dangerous but important job. With this patch we introduce
the concept of a per-service quota. The ds-backup quota defaults to
70% of the disk on which /library/users is, but can be overriden /
controlled from /etc/xs_quota.conf

If the (soft) quota is met, this script will first remove daily
snapshots that are older than its MAX_AGE_DAILY (default: 90 days)
keeping only one monthly snapshot (the earliest of each month).

If after the cleanup of the older dailies we are over quota, it will
remove old snapshots of users that are over the user quota (defined
as ds-backup quota / # of users) until things fit within the quota.

This script has the ability to delete all the user data we are
intending to preserve.