ds-backup: Use rsync -rlt instead of -a -- don't try to set grp or modes!
Using -a a number ofthings can go wrong.

If - accidentally - the user belongs to a group with matching
GID on the XS, the group gets assigned, and all sorts of things go wrong.

The second hunk of this patch will 'touch' the times on the
server side, making sure that incrond triggers postprocessing.