ds-backup.sh: deal with ping not setuid root
OS release 12.1 and 13.1 dropped suid from ping and the olpc user
does not have the privileges to do raw sockets so ping always fails
and ds-backup never completes.  This changes things so we first try to
run it with root privileges using sudo. If a root password has been
set sudo will fail so try one last attempt hoping that ping has been
set suid root. In 13.2 suid root was restored to ping so hopefully the
window of failure will be small.

[dsd: avoid subshell fork for pinging]