ds-backup.py: Add more backup paths
ds-backup.py was single purpose for backing up the datastore but all
the mechanics were in place to backup other things.  This adds 2 paths
to the backup list 'power-logs' and the users documents directory.
The specific purpose of this enhancement was for increasing the ease
of capturing power-logs out in the field.

Now if a XS is in use the power logs will automatically get copied
to the XS.  The documents dir was tossed in for good measure to help
backup things done in GNOME and other stuff the users may do that do
not end up in the journal datastore.